pyproject.toml
[project]
name = "avocado"
version = "0.1.0"
description = "Research project including petri and other phases"
requires-python = ">=3.11"
dependencies = [
    # Install petri as an editable dependency
    "petri",
    # Install tinker-cookbook as an editable dependency
    "tinker-cookbook",

    # Shared dependencies (using petri's newer versions)
    "anthropic>=0.55.0",  # petri needs >=0.55.0 (emergent-values had 0.45.2)
    "openai>=1.91.0",     # petri needs >=1.91.0 (emergent-values had 1.63.0)

    # Inference server dependencies
    "click>=8.0",
    "requests>=2.31",
    "paramiko>=3.0",
    "rich>=13.0",

    # emergent-values dependencies
    "accelerate>=1.1.1",
    "fireworks-ai>=0.15.8",
    "google-generativeai",  # for emergent-values (different from google-genai)
    "huggingface-hub>=0.26.2",
    "inflect>=7.5.0",     # fixed duplicate (was 7.3.1 and 7.5.0)
    "litellm>=1.61.5",
    "networkx>=3.3",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "Pillow>=11.1.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.2",
    "plotly>=5.24.0",
    "dash>=2.18.2",
    "protobuf>=5.29.3",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "torch>=2.0.0",
    "tqdm>=4.67.0",
    "transformers>=4.46.2",
    "vllm>=0.6.6.post1",
]

[project.scripts]
inference-server = "inference_server.inference_server.cli:main"

[project.optional-dependencies]
dev = [
    # Include petri's dev dependencies
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[tool.uv.sources]
# This tells uv to install petri from the local ./external_packages/petri directory in editable mode
petri = { path = "external_packages/petri", editable = true }
tinker-cookbook = { path = "external_packages/tinker-cookbook", editable = true }

[tool.hatch.build.targets.wheel]
# Include both src and inference_server packages
# Note: inference_server has nested structure: inference_server/inference_server/
packages = ["src", "inference_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
